fix: 리뷰내용 반영 (GithubLoginCoordinator)
